+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * [ALFREDAPI-438](https://xenitsupport.jira.com/browse/ALFREDAPI-438): Replace hasReadPermission with hasPermission to avoid an Alfresco bug that prevents usage of that method and throws AcessDeniedException
 * [ALFREDAPI-439](https://xenitsupport.jira.com/browse/ALFREDAPI-439): Fixed issue where category facet values would be displayed with their noderef instead of their name
 
+### Added
+* [ALFREDAPI-438](https://xenitsupport.jira.com/browse/ALFREDAPI-441): Added AspectsToAdd/Remove for creation and copying of nodes. 
+
 ## 2.5.0 (2020-07-15)
 
 ### Added